+++ CHANGELOG.md
@@ -9,6 +9,8 @@
     PMPY-1741 OCEL - Visualization of object graphs (object interaction, object descendants, ...)
 * 15964a428e2e0fc50dcc96570fba39f2e0d40099
     PMPY-1742 - OCEL - Added some filters in the simplified interface
+* fefe740338b702eafb7049a8f777f48f464d157d
+  * PMPY-1953 OCEL - insertion of an attribute pointing to the parent object
 * 88c58a1c6809388be58db2ea6e64fb9b58d913d6
   * PMPY-1763 OCEL - sampling objects
 * 40c741c8298584402cd9ea8e86bcca09f2bfd857
@@ -19,6 +21,8 @@
     PMPY-1812 OCEL - expansion of the set of objects during filtering on object identifiers
 * e2e4d357e2218a48f9b2d6b89690fbbe61cd6906
   * PMPY-1952 OCEL - importing/exporting from/to SQLite
+* 908d0aa843f1a4e9f4a6c69139a3be755c2e8555
+  * PMPY-1951 OCEL - conversion to NetworkX DiGraph
 * e7b4f712e1bd9475c73be5bc82c9a22eb0b745b5
   * PMPY-1764 OCEL - objects interactions summary
 * 43b5976f7040a77c0aa32f5a28c508eb19966fa1